+++ src/main/kotlin/online/ruin_of_future/informative_mc_core/ImcCore.kt
@@ -24,7 +24,7 @@ import net.minecraft.server.MinecraftServer
 import online.ruin_of_future.informative_mc_core.command.ImcCommand
 import online.ruin_of_future.informative_mc_core.config.ModConfig
 import online.ruin_of_future.informative_mc_core.data.ModData
-import online.ruin_of_future.informative_mc_core.token_system.TokenManager
+import online.ruin_of_future.informative_mc_core.auth.TokenManager
 import online.ruin_of_future.informative_mc_core.util.configDir
 import online.ruin_of_future.informative_mc_core.util.gameDir
 import online.ruin_of_future.informative_mc_core.util.getFile
