+++ src/main/kotlin/online/ruin_of_future/informative_mc_core/auth/TokenManager.kt
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/lgpl-3.0.txt>.
  */
-package online.ruin_of_future.informative_mc_core.token_system
+package online.ruin_of_future.informative_mc_core.auth
 
 import java.util.*
 import java.util.concurrent.TimeUnit
