+++ src/main/kotlin/online/ruin_of_future/informative_mc_core/command/Authentication.kt
@@ -19,7 +19,7 @@ import com.mojang.brigadier.builder.LiteralArgumentBuilder
 import net.minecraft.server.command.CommandManager.literal
 import net.minecraft.server.command.ServerCommandSource
 import net.minecraft.text.LiteralText
-import online.ruin_of_future.informative_mc_core.token_system.TokenManager
+import online.ruin_of_future.informative_mc_core.auth.TokenManager
 
 class ImcAuthCommand(
     private val tokenManager: TokenManager
