+++ src/main/kotlin/online/ruin_of_future/informative_mc_core/web_api/ApiServer.kt
@@ -21,7 +21,7 @@ import kotlinx.serialization.ExperimentalSerializationApi
 import online.ruin_of_future.informative_mc_core.config.ModConfig
 import online.ruin_of_future.informative_mc_core.data.ModData
 import online.ruin_of_future.informative_mc_core.tmpDirPath
-import online.ruin_of_future.informative_mc_core.token_system.TokenManager
+import online.ruin_of_future.informative_mc_core.auth.TokenManager
 import online.ruin_of_future.informative_mc_core.util.generateCertificate
 import online.ruin_of_future.informative_mc_core.util.generateKeyPair
 import online.ruin_of_future.informative_mc_core.web_api.handler.*
